Implemented phosolve.py
Phosolve runs the standard cycle of solving. Currently it can't ready the project but that will have to be a future addition.

Also have bug fixes for:
Alpha blending is now possible.
Iteration increment error was fixed by moving the code to a spot that didn't immediately depend on it (race condition).